SVN r4232
Use size_t when dealing with strings and strlen.

+++ src/hardware/serialport/libserial.h
@@ -16,12 +16,13 @@
  *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA.
  */
 
+#include <string.h>
 
 typedef struct _COMPORT *COMPORT;
 
 bool SERIAL_open(const char* portname, COMPORT* port);
 void SERIAL_close(COMPORT port);
-void SERIAL_getErrorString(char* buffer, int length);
+void SERIAL_getErrorString(char* buffer, size_t length);
 
 #define SERIAL_1STOP 1
 #define SERIAL_2STOP 2
